Add reposition to taxon
