+++ api/app/controllers/spree/api/v2/platform/menu_items_controller.rb
@@ -3,18 +3,7 @@ module Spree
     module V2
       module Platform
         class MenuItemsController < ResourceController
-          def reposition
-            spree_authorize! :update, resource if spree_current_user.present?
-
-            @new_parent = scope.find(permitted_resource_params[:new_parent_id])
-            new_index = permitted_resource_params[:new_position_idx].to_i
-
-            if resource.move_to_child_with_index(@new_parent, new_index)
-              render_serialized_payload { serialize_resource(resource) }
-            else
-              render_error_payload(resource.errors)
-            end
-          end
+          include ::Spree::Api::V2::Platform::NestedSetRepositionHelper
 
           private
