+++ api/app/controllers/spree/api/v2/platform/taxonomies_controller.rb
@@ -1,0 +1,16 @@
+module Spree
+  module Api
+    module V2
+      module Platform
+        class TaxonomiesController < ResourceController
+
+          private
+
+          def model_class
+            Spree::Taxonomy
+          end
+        end
+      end
+    end
+  end
+end
